refactor(ui): adopt dotted filename convention for mosaic org sections
Rename the delete/leave section slices to <name>.<layer> (.machine/.controller/.view),
update the wrappers, swingset stories, and the mosaic-architecture file-shape doc
(also refreshing the stale wrapper example to render null until ready).
